new cr

+++ dsmg/public/packages/script/paneltesting.js
@@ -1,11 +1,46 @@
 window.onload = function(){
 	$('#paneltesting').addClass('active');
-	$('#inspection_date').datepicker({dateFormat: 'dd-mm-yy', maxDate: "+0D"});
-	$("#inspection_date").datepicker("setDate", new Date());
+	$('#testing_date').datepicker({dateFormat: 'dd-mm-yy', maxDate: "+0D"});
+	$("#testing_date").datepicker("setDate", new Date());
 allPaneltestingstnlcgaten();
+allsupervisors();
+allsupervisorsdesig();
+init();
+loadPanelTestingLedger();
+}
+
+
+function init(){
+	$('#designation').empty();
+	$('#designation').append('<option value="DSTE/TSK">DSTE/TSK</option>');
+	$('#designation').append('<option value="ADSTE/MXN">ADSTE/MXN</option>');
+	$('#designation').append('<option value="ADSTE/TSK">ADSTE/TSK</option>');
+	$('#maintenance_by').empty().append('<option value="NA">--NA--</option>');
 }
 var token =  $("input[name=_token]").val();
 
+
+function saveData(){
+	var formData = $('form#paneltesting-form').serializeArray();
+	$('#saveBtn').attr('disabled',true).html('<i class="fa fa-spinner fa-spin"></i>');
+		 $.ajax({
+			url:'/schedule/savepaneltesting',
+			type: 'POST',
+			dataTtype: 'JSON',
+			data: formData,
+			success: function(data){
+				console.log(data)
+				$('#saveBtn').attr('disabled',false).html('save');
+				if(data!=0){
+					alertify.success('data saved successfully');
+					loadPanelTestingLedger();
+				}
+				else{
+					alertify.error('Not able to save at this moment.Please contact administrator!');
+				}
+			}
+		});
+}
 function allPaneltestingstnlcgaten(){
 	$.ajax({
 		url: '/common/paneltestingstnlcgate',
@@ -14,10 +49,28 @@ function allPaneltestingstnlcgaten(){
 		success: function(data){
 			//$('#station_id').append('<option value="0">--station code--</option>');
 			for (var i in data){
-				$('#station_id').append('<option value="'+data[i].id+'">'+data[i].code+'</option>');
+				$('#station_id').append('<option value="'+data[i].id+'">'+data[i].stn_lc_gate+'</option>');
 			}
 		}
 	});
+}
+var supervisors = new Array();
+var supervisors_desig = new Array();
+
+function choosingRole(){
+	var role =$("#role").val();
+	if(role=="Officer"){
+		$('#designation').empty();
+	 	$('#designation').append('<option value="DSTE/TSK">DSTE/TSK</option>');
+		$('#designation').append('<option value="ADSTE/MXN">ADSTE/MXN</option>');
+		$('#designation').append('<option value="ADSTE/TSK">ADSTE/TSK</option>');
+		$('#maintenance_by').empty().append('<option value="NA">--NA--</option>');
+
+	}else if (role=="Supervisor") {
+				showSupervisorDesig(supervisors_desig);
+				showSupervisor(supervisors);
+	}
+
 }
 
 function allsupervisors(){
@@ -27,12 +80,24 @@ function allsupervisors(){
 	 type: 'GET',
 	 dataTtype: 'JSON',
 	 success: function(data){
-		 $('#maintenance_by').append('<option value="NA">--NA--</option>');
-		 for(var i in data){
-			 $('#maintenance_by').append('<option value="'+data[i].name+'">'+data[i].name+'</option>');
-		 }
+		 supervisors=data;
+
 	 }
  });
+}
+function showSupervisor(data){
+	$('#maintenance_by').empty();
+	$('#maintenance_by').append('<option value="NA">--NA--</option>');
+	for(var i in data){
+		$('#maintenance_by').append('<option value="'+data[i].name+'">'+data[i].name+'</option>');
+	}
+}
+function showSupervisorDesig(data){
+	$('#designation').empty();
+	$('#designation').append('<option value="NA">--NA--</option>');
+  for(var i in data){
+ 	 $('#designation').append('<option value="'+data[i].name+'">'+data[i].name+'</option>');
+  }
 }
 function allsupervisorsdesig(){
 	$('#designation').empty();
@@ -41,45 +106,26 @@ function allsupervisorsdesig(){
 	 type: 'GET',
 	 dataTtype: 'JSON',
 	 success: function(data){
-		 $('#designation').append('<option value="NA">--NA--</option>');
-		 for(var i in data){
-			 $('#designation').append('<option value="'+data[i].name+'">'+data[i].name+'</option>');
-		 }
+		 supervisors_desig=data;
+
 	 }
  });
 }
-function loadCrossingPointInspectionLedger(){
+function loadPanelTestingLedger(){
 	$('#overdueBtn').attr('onclick','overdueStation()').attr('class','btn btn-danger btn-block').html('overdue station');
 
-	$('#table_level').empty().html('<h5><span class="label label-success">Joint Point & Crossing Inspection Ledger</span></h5>');
-	$('#table_header').empty().html('<tr><th>Station Code</th><th>Role</th><th>Designation</th><th>Maintenance by</th><th>Last Inspection Date</th><th>Next Inspection Date Due</th><th>Delete</th></tr>');
-	$('#data-list').html('<tr><td colspan="9"><center><i class="fa fa-spinner fa-spin fa-3x"></i></center></td></tr>')
+	$('#table_level').empty().html('<h5><span class="label label-success">Panel Testing Ledger</span></h5>');
+	$('#table_header').empty().html('<tr><th>Station/LC Gate</th><th>Role</th><th>Designation</th><th>Maintenance by</th><th>Last testing date</th><th>Next testing date</th><th>Delete</th></tr>');
+	$('#data_list').html('<tr><td colspan="9"><center><i class="fa fa-spinner fa-spin fa-3x"></i></center></td></tr>')
 
 	$.ajax({
-		url: '/schedule/allcrossinginspectionledger',
+		url: '/schedule/allpaneltestingledger',
 		type: 'GET',
 		dataTtype: 'JSON',
 		success: function(data){
-			$('#data-list').empty();
+			$('#data_list').empty();
 			 for (var i in data){
-
-				 var today=moment(new Date()).format('DD/MM/YY');
-
-				 var next_date_row=moment(data[i].due_date_of_inspection).format('DD/MM/YY');
-
-
-				 if(data[i].role=='SS'){
-				 	next_date_row=next_date_row+'  for IC';
-				}else{
-					next_date_row=next_date_row+'  for SS';
-				}
-
-				 if(moment(data[i].due_date_of_inspection).isAfter(new Date().getTime())){
-				 			next_date_row='<td >'+next_date_row+'</td>';
-					 }else{
-						 next_date_row='<td class="text text-danger">'+next_date_row+'</td>';
-
-					 }
+				 var next_date_row='<td>'+moment(data[i].testing_date).format('DD/MM/YY')+'</td>';
 
 				var row = '<tr>'
 						+'<td class="hidden id">'+data[i].id+'</td>'
@@ -87,11 +133,11 @@ function loadCrossingPointInspectionLedger(){
 						+'<td>'+data[i].role+'</td>'
 						+'<td>'+data[i].designation+'</td>'
 						+'<td>'+data[i].maintenance_by+'</td>'
-						+'<td>'+moment(data[i].date_of_inspection).format('DD/MM/YY')+'</td>'
+						+'<td>'+moment(data[i].testing_date).format('DD/MM/YY')+'</td>'
 						+next_date_row
 						+'<td><button class="del btn btn-rounded btn-sm btn-icon btn-danger"><i class="fa fa-trash"></i></button></td>'
 						+'</tr>';
-				$('#data-list').append(row);
+				$('#data_list').append(row);
 			}
 		}
 	});
@@ -99,15 +145,17 @@ function loadCrossingPointInspectionLedger(){
 }
 
 //delete data
-$("#data-list").on("click", ".del", function(){
+$("#data_list").on("click", ".del", function(){
 	$deleting = $(this);
 	var id = $deleting.closest("tr").find(".id").text();
-	$deleting.html('&nbsp;&nbsp;<i class="fa fa-spinner fa-spin fa-lg"></i>&nbsp;&nbsp;');
-	delete_data(id);
+	alertify.confirm('Please confirm your delete!').set('onok', function(closeEvent){
+				$deleting.html('&nbsp;&nbsp;<i class="fa fa-spinner fa-spin fa-lg"></i>&nbsp;&nbsp;');
+				delete_data(id);
+	});
 });
 function delete_data(id){
 	$.ajax({
-		url: '/schedule/deletecrossinginspection/'+id,
+		url: '/schedule/deletepaneltestingledger/'+id,
 		type: 'GET',
 		dataType: 'JSON',
 		success: function(data){
